production build, codestyle fixes
