+++ README.md
@@ -1,5 +1,7 @@
-Программа запускается через исполняемый файл "main".
-
 Цель программы: "наложить" матрицу вёрстки на изображение, вызвав тем самым изменение его резкости.  
 Программа поддерживает bmp изображения с 24 и 32 битной кодировкой пикселя.  
 Матрицу вёрстки можно задать файлом, в начале которого пишутся её размеры, или тремя числами для матрицы 3x3.
+
+Сборка производится через команду
+bash ./src/build.sh
+Собранный дистрибутив поставляется через директорию build
